backup/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.15
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.0.6
langchain-text-splitters==0.0.2
langsmith==0.1.147
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.107.3
orjson==3.11.3
packaging==23.2
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-telegram-bot==20.7
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==8.5.0
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
